upgrading to vue3
Upgrading vue and other devDependencies to the latest versions
