+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "vuewordcloud",
-	"version": "18.7.11",
+	"version": "19.0.0",
 	"description": "Generates a cloud out of the words.",
 	"keywords": [
 		"cloud",
@@ -22,16 +22,17 @@
 		"serve": "rollup -c -w"
 	},
 	"devDependencies": {
-		"@rollup/plugin-buble": "^0.21.3",
-		"@rollup/plugin-node-resolve": "^8.0.0",
-		"eslint": "^7.1.0",
-		"eslint-config-seregpie": "^1.0.1",
-		"rollup": "^2.12.0",
-		"rollup-plugin-serve": "^1.0.1",
-		"rollup-plugin-stringify": "^19.0.2",
-		"rollup-plugin-terser": "^6.1.0"
+		"@rollup/plugin-buble": "^1.0.1",
+		"@rollup/plugin-node-resolve": "^15.0.1",
+		"@rollup/plugin-terser": "^0.1.0",
+		"eslint": "^8.28.0",
+		"eslint-config-seregpie": "^1.2.1",
+		"rollup": "^3.5.0",
+		"rollup-plugin-serve": "^2.0.2"
+	},
+	"dependencies": {
 	},
 	"peerDependencies": {
-		"vue": "^2.6.11"
+		"vue": "^3.x"
 	}
 }
